slack-jira-bridge/requirements.txt
slack-bolt==1.20.1
jira==3.5.2
python-dotenv==1.0.1
PyYAML==6.0
requests==2.32.3
